Add Null Check for SubworldSystem.Current
This change allows the "Current" Mod call to properly comply to the documented behavior. Currently, if the user is not in a subworld, SubworldSystem.Current will be null, causing attempting to access SubworldSystem.Current.FullName to throw.

This change adds the null coalescence operator, causing the returned value to be null if SubworldSystem.Current is null, complying to the documented behavior and preventing a throw if this call is used when the user is not in a subworld.